+++ packages/plugins/FileService/src/SNSAdaptor/modals/modals.ts
@@ -1,4 +1,4 @@
 import { SingletonModal } from '@masknet/shared-base'
-import type { ConfirmModalOpenProps } from './Confirm/index.js'
+import type { ConfirmModalOpenProps } from './ConfirmModal/index.js'
 
 export const ConfirmModal = new SingletonModal<ConfirmModalOpenProps>()
